+++ packages/hpe-design-tokens/package.json
@@ -25,7 +25,7 @@
   "homepage": "https://design-system.hpe.design/",
   "scripts": {
     "prettier:check": "prettier --check src/",
-    "figma-api-cli": "tsx src/scripts/figma-api-cli/index.ts",
+    "figma-variables-cli": "tsx src/scripts/figma-variables-cli/index.ts",
     "sync-tokens-to-figma": "tsx src/scripts/sync_tokens_to_figma.ts",
     "sync-figma-to-tokens": "tsx src/scripts/sync_figma_to_tokens.ts",
     "build-style-dictionary": "tsx src/scripts/build-style-dictionary.js",
