+++ packages/hpe-design-tokens/src/scripts/figma-variables-cli/__tests__/fixtures.ts
