+++ packages/hpe-design-tokens/src/scripts/figma-variables-cli/__tests__/get-variable-by-id.test.ts
