+++ packages/hpe-design-tokens/src/scripts/figma-variables-cli/__tests__/options.test.ts
