+++ packages/hpe-design-tokens/src/scripts/figma-variables-cli/figma-variables-cli.ts
