+++ packages/hpe-design-tokens/src/scripts/figma-variables-cli/get-modes.ts
