+++ packages/hpe-design-tokens/src/scripts/figma-variables-cli/get-variable-by-id.ts
