+++ packages/hpe-design-tokens/src/scripts/figma-variables-cli/get-variables.ts
