+++ packages/hpe-design-tokens/src/scripts/figma-variables-cli/index.ts
@@ -1,4 +1,4 @@
-import cli from './figma-api-cli.js';
+import cli from './figma-variables-cli.js';
 import { brightRed } from '../../utils.js';
 
 cli().catch(error => {
