+++ packages/hpe-design-tokens/src/scripts/figma-variables-cli/output.ts
