+++ packages/hpe-design-tokens/src/scripts/figma-variables-cli/payload.ts
