+++ packages/hpe-design-tokens/src/scripts/figma-variables-cli/post-variables.ts
