+++ packages/hpe-design-tokens/src/scripts/figma-variables-cli/types.ts
